Move book class and book methods to module book.js
